create update and check func

+++ app/remove.php
@@ -1,24 +1,22 @@
 <?php
 
-if (isset($_POST['id'])){
+if ($_SERVER['REQUEST_METHOD'] === 'POST' && !empty($_POST['id'])) {
     require '../db_conn.php';
 
     $id = $_POST['id'];
 
-    if (empty($id)){
-        echo 0;
-    }else{
-        $stmt = $conn->prepare("DELETE FROM todos WHERE id=?");
-        $res = $stmt->execute([$id]);
-        if ($res){
-            echo 1;
-        }else{
-            echo 0;
-        }
-        $conn = null;
-        exit();
-    }
-}else{
-    header("Location: ../index.php?mess=error");
-}
+    $stmt = $conn->prepare("DELETE FROM todos WHERE id=?");
+    $res = $stmt->execute([$id]);
 
+    if ($res) {
+        echo 'Success';
+    } else {
+        echo 'Error: Could not delete item';
+    }
+
+    $conn = null;
+} else {
+    header("Location: ../index.php?mess=error");
+    exit();
+}
+?>
